Changing threading and added thread manager
Splitted thread_base.py into 3 files. One for the thread manager base,
the other for the interface, and the last one for the agent thread
manager and DQN thread.

The class BaseThreadManager is used to provide typing support for the
class BaseTrainingAgent. The variables in BaseTrainingAgent has moved
to __init__() in order to be more pythonic??? (Not sure if it is)

Lastly .gitignore and run_development.py are used for development
purposes. For launching only the interface: run_development.py