Functions and files reordering + disabled secondary selector and panel + removed ball control panel
All functions that are returning functions now have
build as prefix. The functions that are producing
tkinter elements now have the make as prefix.

Removed the ball control panel as it is not very useful
for now. Disabled secondary selector and its panel
because it is very buggy. This feature will come back.